fix/시간 표시용 컴포넌트 createdTime.js 생성, 해당되는 페이지 최적화
