+++ src/pages/board/CommentItem.jsx
@@ -1,3 +1,4 @@
+import createdTime from "@pages/board/createdTime";
 import useUserStore from "@zustand/useUserStore";
 import PropTypes from "prop-types";
 
@@ -9,34 +10,19 @@ export default function CommentItem({ item }) {
   const { user } = useUserStore();
   console.log(item);
 
-  const createdTime = (createdDate) => {
-    const formatRelativeTime = (inputDate) => {
-      const now = new Date();
-      const pastDate = new Date(inputDate);
-      const minDiff = Math.floor((now - pastDate) / (1000 * 60));
-      if (minDiff < 1) return "방금 전";
-      if (minDiff < 60) return `${minDiff}분 전`;
-      if (minDiff < 1440) return `${parseInt(minDiff / 60)}시간 전`;
-      if (minDiff < 2880) return `${parseInt(minDiff / 1440)}일 전`;
-
-      // 이틀 이상인 경우에는 날짜를 표시
-      return pastDate.toLocaleString();
-    };
-
-    return formatRelativeTime(createdDate);
-  };
-
   const newDate = createdTime(item.createdAt);
   return (
     <>
-      <div className="flex flex-row mt-5 px-[15px]">
+      <div className="flex flex-row mt-5 px-[15px] items-center">
         <img
           src={`https://11.fesp.shop${item.user.image}`}
           alt="ProfileImage"
           className="w-6 h-6 rounded-full object-cover"
         />
         <span className="mx-[5px] text-sm">{item.user.name}</span>
-        <span className="text-[10px] ml-auto text-gray4">{newDate}</span>
+        <span className="text-[10px] ml-auto text-gray4 self-start">
+          {newDate}
+        </span>
       </div>
       <div className="flex pb-5 border-b-[1px] border-gray3/50 px-[15px]">
         <div className="mt-3 text-xs text-gray5 pl-5">{item.content}</div>
